Breakup the `build_env` module into smaller modules
This separates the various concerns internally, while still exposing the
same API for the sub-package.
